added class for common methods
